separate out segment averaging from entire trip
